Rename AircloakApi namespace -> Aircloak and remove it from Explorer.Api namespace
